get claims from token

+++ TweetBook/Controllers/V1/PostsController.cs
@@ -7,6 +7,7 @@ using TweetBook.Contracts.V1;
 using TweetBook.Contracts.V1.Request;
 using TweetBook.Contracts.V1.Responses;
 using TweetBook.Domain;
+using TweetBook.Extensions;
 using TweetBook.Services;
 
 namespace TweetBook.Controllers.V1
@@ -49,7 +50,8 @@ namespace TweetBook.Controllers.V1
         {
             var post = new Post
             {
-                Name = postRequest.Name
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
             };
 
             if (post.Id != Guid.Empty)
@@ -72,11 +74,19 @@ namespace TweetBook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Updated([FromRoute] Guid postId, [FromBody] UpdatePostRequest request)
         {
-            var post = new Post
+            var userOwnPost = await _postService.UserOwnPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnPost)
             {
-                Id = postId,
-                Name = request.Name
-            };
+                return BadRequest(new
+                {
+                    error = "You do not own this post"
+                });
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+
+            post.Name = request.Name;
 
             var updated = await _postService.UpdatePostAsync(post);
 
@@ -89,6 +99,16 @@ namespace TweetBook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnPost = await _postService.UserOwnPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnPost)
+            {
+                return BadRequest(new
+                {
+                    error = "You do not own this post"
+                });
+            }
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
